OS Error related checks and "DCAF:" added to errors.
Adding new error checks in case you specify the wrong operating system.
Using the manual deploy tool I checked to see the errors we got if you
specified the wrong operating system but otherwise were correctly
configured. I added in some extra error checking and words to our errors
to recommend checking that the operating system is correct.

The OS check was inspired by this forum post:
http://forums.ni.com/t5/Distributed-Control-Automation/1-1-Release-Notes-for-DCAF-Core/gpm-p/3616864#M197

I also added "DCAF:" to the appended errors we show to the user.