Update MCLB Refresh
Updated MCLB Refresh to use the active cell row instead of value when
setting the value. With this change, the highlighted row will not reset
to the first row when moving between modules. Somewhat of a cosmetic fix
but it had potential for the active cell to not be on a highlighted row
which could result in a few corner case UI quirks.

For example, user highlights row 2 of modbus module and clicks on scan
engine module. Now when the user selects modbus module again it will
stay highlighted on row 2.