Change Handling of Left-Click Discards
Cleaned up Left Click.vi and changed how VI determined whether event was
going to be discarded.

Previously, discard was false when the column was read only, allowing
users to edit read-only cells form the MCLB. Now, when the column is
read only, the VI checks whether the user selected cell is in the
highlighted row and will discard if they are equal.

This implementation will cause events to be incorrectly discarded if the
user is editing a cell and tries to push the edits by clicking on a
read-only cell in the same row.